Backend/requirements.txt
asgiref==3.8.1
dj-rest-auth==7.0.0
Django==5.1.1
django-allauth==65.2.0
django-cors-headers==4.6.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
PyJWT==2.10.0
python-decouple==3.8
sqlparse==0.5.1